# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "bidding_app",
    "description": "A simple bidding app",
    "contact": {
      "name": "Tom Moulard",
      "email": "tom@moulard.org"
    },
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "paths": {
    "/login": {
      "post": {
        "tags": [],
        "operationId": "login_user",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Login a user",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/refresh": {
      "post": {
        "tags": [],
        "operationId": "refresh_token",
        "requestBody": {
          "content": {
            "text/plain": {
              "schema": {
                "type": "string"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Refresh JWT token",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/register": {
      "post": {
        "tags": [],
        "operationId": "register_user",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Register a new user",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/purchase": {
      "post": {
        "tags": [],
        "operationId": "purchase_points",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "number",
                "format": "float"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Purchase points",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    },
                    "points": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "User": {
        "type": "object",
        "required": [
          "email",
          "username",
          "password"
        ],
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "username": {
            "type": "string"
          },
          "points": {
            "type": "integer"
          }
        }
      }
    }
  },
  "tags": [
    {
      "name": "user",
      "description": "User management endpoints"
    }
  ]
}